Restore UART0 code, add FPU support, patch panics
The code for UART1 is moved to mini_uart.rs, and the code for UART0 is
restored in uart.rs. This is because UART1 is not emulated properly by
Qemu.

The start code is now located at 0x0 and jumps to 0x8000.
The kernel depends on compiler_builtins to avoid missing intrinsics.

FPU is enabled at kernel start.

The custom panic handler should now work.

Known bugs:
Using the write! macro or core::fmt::write cause a branch to 0x0.