First try to make an higher half kernel
This code do not work as it is. There is still some code that is calling
functions in the lower half (and it makes the kernel crash).

In this commit I wanted to clearly divide physical and virtual addresses
and I created different types to distinguish them.

I think that this version is really more ugly than the lower half one.
And using GDB with it is awful.